# Gesture typing improvements