--- a/app/src/main/res/values/dimens.xml
+++ b/app/src/main/res/values/dimens.xml
@@ -94,7 +94,7 @@
     <dimen name="quick_key_size">50dp</dimen>
 
     <dimen name="yochees_radius_key_corner">5dp</dimen>
-    <dimen name="gesture_typing_curvature">5sp</dimen>
+    <dimen name="gesture_typing_min_point_distance">5sp</dimen>
     <dimen name="gesture_stroke_width">4dp</dimen>
     <dimen name="watermark_size">12sp</dimen>
     <dimen name="watermark_margin">3dp</dimen>