This pull request greatly improves the speed+quality of gesture typing, at least for me!
In particular:

- **Faster path generation**: From 12s to 3.5s (English, debug mode), from 17s to 3s (German, debug mode)
- **Faster candidate finding**: From 120ms to 80ms when typing "implementation" on an English Keyboard (in debug mode)
- **Better special character handling**: Previously, when a keyboard had two keys with the same base character (e.g., "ä" and "a" on the German keyboard), the path generation would always choose the base character, making it nearly impossible to gesture type words like "Männer". This now works properly.
- Most importantly: **Much better word recognition**: At least for me, the recognition of the gesture typed words is much better, especially for long words. For this, I mainly reduced the number of points that are chosen for the path gesture-typed by the user. The points on this path are now almost only the ones where the path changes direction, leading to a higher accuracy.

Some more details:
- I split up the variable `mCurvatureSize`. Previously, it was used with two meanings: as the minimum distance between two points, and as the number of points to the left/right used for curvature calculation. We now have `CURVATURE_NEIGHBORHOOD` and `mMinPointDistance`.
- Most of the speedup also comes from a `mKeysByCharacter` lookup, which is now used. Previously, we iterated over all keys to check whether they matched the character we're looking for.
- I added some more debug logging for easier measuring of the performance.