@menny 2019-04-15
Can I mention you in the changelog? In regard to this improvement?