@topas-rec 2019-05-27
I just tested it by typing this comment.  IT IS AMAZING! Many thanks!